Make self.env.vars an AttributedDict
This is to have the "original" vars object available in python commands.
We do not want to copy it, because we then just can alter the copy, but
not the "real" values. This allows us both: Altering the values and
using the attribute notation values. This allows us both: Altering the
values and using the attribute notation.
